image-shrink/requirements.txt
dify_plugin~=0.0.1b67
pillow